Widen section labels on swissen template
I noticed that on the swissen template, the word "experience" was a bit
too long to use as a section label. Given that this is a pretty short
word, and also a very common section label for résumés, I widened the
section labels from 16% of the total width to 20% (and reduced the width
of the main text area to match) for both html and pdf renderings.